### 2016-08-16 #52 @whedon
[REVIEW]: fuse: An R package for ensemble Hydrological Modelling
**Submitting author:** @cvitolo (<a href="http://orcid.org/0000-0002-4252-1176">Claudia Vitolo</a>)
**Repository:** <a href="https://github.com/cvitolo/fuse" target ="_blank">https://github.com/cvitolo/fuse</a>
**Version:** v3.0.0
