### 2016-08-21 #52 @pragyansmita
> library(devtools)
> install_github("cvitolo/r_fuse", subdir = "fuse")
Downloading GitHub repo cvitolo/r_fuse@master
from URL https://api.github.com/repos/cvitolo/r_fuse/zipball/master
Error: Could not find build tools necessary to build fuse

I am using R 3.2.5 and have Rtools33 installed. I am still receiving the above error. Any idea how to move beyond to install fuse. Thanks!
