### 2016-08-26 #52 @cvitolo
@pragyansmita I'm sorry you cannot install fuse. I work only with Linux machines and I cannot reproduce your problem. I use travis-ci to check that there are no problems with installation and tests. As you can see [here](https://travis-ci.org/cvitolo/r_fuse) the package passes all the tests and builds no problem on a linux machine.
