### 2016-08-26 #52 @masalmon
@cvitolo by the way on Travis you can also test on OSX by adding

```
os:
 - linux
 - osx
```

after ` language: r` at the top of your .travis.yml.

Then you can use Appveyor for continuous integration on Windows (there's a `devtools::use_appveyor()` function which automatically creates the Appveyor config file, and then you need to activate CI for the repository on [Appveyor website](https://www.appveyor.com/) ), quite similarly to what you did on Travis.

Furthermore on Travis and Appveyor there are options for testing your package with several R versions, I do it [here](https://github.com/ropenscilabs/monkeylearn/blob/master/appveyor.yml) and [here](https://github.com/ropenscilabs/monkeylearn/blob/master/.travis.yml).

Not sure if it'll help you see @pragyansmita's problem, but who knows :-)
