### 2016-08-26 #52 @cvitolo
Many thanks @masalmon!

I just added the os: linux osx lines to .travis.yml. Building the job for osx takes ages, is that normal?

Also tried to configure appveyor but I keep getting this error 'Specify a project or solution file. The directory does not contain a project or solution file.'... not sure how to fix this.
