### 2016-08-26 #52 @masalmon
You're welcome!

I don't know for osx, maybe it's just because it's the first build?

I am sorry I can't help for Appveyor, I never had this issue. Did you need to do something specific on Travis because of the sub directory structure?
