### 2016-08-26 #52 @cvitolo
I just did:

cd fuse

I tried the same in appveyor but that did not fix the problem
