### 2016-08-26 #52 @masalmon
Maybe this http://stackoverflow.com/questions/33411765/check-multiple-subdirectories-using-appveyor can help?
