### 2016-08-26 #52 @janschulz
> Specify a project or solution file. The directory does not contain a project or solution file.

I thought this happens when appveyor doesn't see a appveyor.yml file or the file doesn't have a `build` section (or whatever the section is named which builds the project) -> per default it tries to build a MS Visual Studio solution / project file (like the default makefile for make based systems).
