### 2016-08-26 #52 @jkahn
Apologies for intervening, but could these details be discussed in an issue on the submitted project rather than here in the review?

I'll file a ticket with @arfon to encourage this as a general directive to reviewers.
