### 2016-08-26 #52 @arfon
> Apologies for intervening, but could these details be discussed in an issue on the submitted project rather than here in the review?

Good suggestion @jkahn. @masalmon @cvitolo @janschulz - would you mind taking this discussion over to the submission repo (https://github.com/cvitolo/r_fuse)?
