### 2016-09-04 #52 @cvitolo
Thanks @arfon! I opened an issue on the r_fuse repo. We will continue the discussion there.
