+++ README.md
@@ -1,0 +1,2 @@
+# fuse
+fuse: Framework for Understanding Structural Errors (Hydrological Modelling)
