+++ src/fuse.cpp
@@ -1,0 +1,1836 @@
+// Use CRAN package BH for Boost headers
+
+// [[fuse::depends(BH)]]
+#include "fuse.h"
+#include "modelStructure.h"
+#include "smaStructures.h"
+
+#include <boost/math/distributions/gamma.hpp>
+#include <boost/numeric/odeint.hpp>
+
+using boost::math::gamma_distribution;
+using boost::math::cdf;
+using boost::math::isnan;
+
+// Global pointers for ODE solver
+SmaInputParameters* globalInputParams = 0;
+SmaDerivedParameters* globalDerivedParams = 0;
+double* globalPrecipitation = 0;
+int globalPrecipitationLength = 0;
+double* globalPotentialEvapotrans = 0;
+int globalPotentialEvapotransLength = 0;
+ModelStructure* globalSmodl = 0;
+int globalTimeCounter = 0;
+double* globalSavedStates = 0;
+bool globalWriteSavedStates = false;
+
+struct streaming_observer
+{
+    std::ostream &m_out;
+    streaming_observer( std::ostream &out ) : m_out( out ) {}
+
+    void operator()( const state_type &x , double t ) const
+    {
+        // m_out << "In observer at t=" << t << endl;
+        if(t > 0){
+            upstates((state_type&)x);
+            if(globalWriteSavedStates){
+                for(int j=0; j<10; j++){
+                    globalSavedStates[globalTimeCounter*10 + j] = x[j];
+                }
+            }
+            globalTimeCounter += 1;
+        }
+
+        //m_out << t;
+        //for( size_t i=0 ; i < x.size() ; ++i )
+        //    m_out << "\t" << x[i];
+        //m_out << "\n";
+    }
+};
+
+
+
+
+bool fuseRoutingSim( double* instantaneousDischarge, int instantaneousDischargeSize, int* modelParams, int modelParamsSize, double timeDelay, double timeStep, double* routedDischarge, int routedDischargeSize ){
+
+    if(modelParamsSize != 8){
+        // Looks like the mode parameters are not sensible
+        return false;
+    }
+    
+    ModelStructure ms;
+    
+    ms.rainfallError        = (RainfallErrorType)modelParams[0]; // rferr
+    ms.upperSoilLayerArch   = (UpperSoilLayerArchType)modelParams[1]; // arch1
+    ms.lowerSoilLayerArch   = (LowerSoilLayerArchType)modelParams[2]; // arch2
+    ms.runoff               = (RunoffType)modelParams[3]; // qsurf
+    ms.percolation          = (PercolationType)modelParams[4]; // qperc
+    ms.evaporation          = (EvaporationType)modelParams[5]; // esoil
+    ms.interflows           = (InterflowsType)modelParams[6]; // qintf
+    ms.routing              = (RoutingType)modelParams[7]; // q_tdh
+
+    /*
+    bool debug = true;
+    if(debug){
+        cout << endl << endl << "fuseRoutingSim called with the following parameters:" << endl;
+        cout << "instantaneousDischarge at " << &instantaneousDischarge[0] << endl;
+        cout << "instantaneousDischargeSize" << " : " << instantaneousDischargeSize << endl;
+        cout << "modelId" << " : " << modelId << endl;
+        cout << "modelListFileName at " << &(modelListFileName[0]) << endl;
+        cout << "modelListFileName" << " : " << modelListFileName << endl;
+        cout << "modelListFileName[0]" << " : " << modelListFileName[0] << endl;
+        cout << "timeDelay" << " : " << timeDelay << endl;
+        cout << "timeStep" << " : " << timeStep << endl;
+        cout << "routedDischarge at " << &routedDischarge[0] << endl;
+        cout << "routedDischargeSize" << " : " << routedDischargeSize << endl;
+    }
+    */
+
[…]
