### 2016-09-05 e637c2d
Added travis and updated README
