+++ paper/paper.bib
@@ -1,0 +1,26 @@
+@Manual{fuse-archive,
+  author       = {Claudia Vitolo and Peter Wells and Martin Dobias and Wouter Buytaert},
+  title        = {fuse: Framework for Understanding Structural Errors, R package version 3.0.0},
+  year         = 2012,
+  doi          = {10.5281/zenodo.14005},
+  url          = {https://github.com/cvitolo/r_fuse}
+}
+
+@Manual{R-base,
+    title = {R: A Language and Environment for Statistical Computing},
+    author = {{R Core Team}},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    year = {2016},
+    url = {https://www.R-project.org/},
+}
+
+@article{Clark2008,
+  title = {Framework for Understanding Structural Errors (FUSE): A modular framework to diagnose differences between hydrological models},
+  author = {Martyn P. Clark and Andrew G. Slater and David E. Rupp and Ross A. Woods and Jasper A. Vrugt and Hoshin V. Gupta and Thorsten Wagener and Lauren E. Hay},
+  year = {2008},
+  journal = {Water Resour. Res},
+  volume = {44},
+  number = {W00B02},
+  doi = {10.1029/2007WR006735},
+}
