+++ README.md
@@ -1,6 +1,3 @@
-# README for the fuse package
-Claudia Vitolo  
-`r Sys.Date()`  
 
 
 
@@ -10,9 +7,9 @@ fuse: Framework for Understanding Structural Errors (Hydrological Modelling)
 ======
 
 [![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.14005.svg)](http://dx.doi.org/10.5281/zenodo.14005)
-[![Build Status](https://travis-ci.org/cvitolo/r_fuse.svg)](https://travis-ci.org/cvitolo/r_fuse.svg?branch=master)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/cvitolo/r_fuse?branch=master&svg=true)](https://ci.appveyor.com/project/cvitolo/r_fuse)
-[![codecov.io](https://codecov.io/github/cvitolo/r_fuse/coverage.svg?branch=master)](https://codecov.io/github/cvitolo/r_fuse?branch=master)
+[![Build Status](https://travis-ci.org/cvitolo/fuse.svg)](https://travis-ci.org/cvitolo/fuse.svg?branch=master)
+[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/cvitolo/fuse?branch=master&svg=true)](https://ci.appveyor.com/project/cvitolo/fuse)
+[![codecov.io](https://codecov.io/github/cvitolo/fuse/coverage.svg?branch=master)](https://codecov.io/github/cvitolo/fuse?branch=master)
 
 Implementation of the framework for hydrological modelling FUSE described in Clark et al. (2008) and based on the Fortran code provided by M. Clark in 2011. The package consists of two modules: Soil Moisture Accounting module (fusesma.sim) and Gamma routing module (fuserouting.sim). It also contains default parameter ranges (fusesma.ranges and fuserouting.ranges) and three data objects: DATA (sample input dataset), parameters (sample parameters) and modlist (list of FUSE model structures).
 
@@ -34,8 +31,7 @@ You can install this package from Github with [devtools](https://github.com/hadl
 
 
 ```r
-library(devtools)
-install_github("cvitolo/r_fuse", subdir = "fuse")
+devtools::install_github("cvitolo/fuse")
 ```
 
 Load the rdefra package:
@@ -240,6 +236,6 @@ summary(modfit)
 
 ## Meta
 
-* Please [report any issues or bugs](https://github.com/cvitolo/r_fuse/issues).
+* Please [report any issues or bugs](https://github.com/cvitolo/fuse/issues).
 * License: [GPL-3](https://opensource.org/licenses/GPL-3.0)
 * Get citation information for the `fuse` package in R doing `citation(package = 'fuse')`
