+++ .travis.yml
@@ -1,0 +1,13 @@
+language: r
+os:
+ - linux
+
+cache: packages
+before_install:
+  - cd fuse
+
+r_github_packages:
+  - jimhester/covr
+
+after_success:
+  - Rscript -e 'covr::codecov()'
