### 2016-09-05 4f0a02e
amended travis to build on the root dir
