+++ .travis.yml
@@ -3,8 +3,6 @@ os:
  - linux
 
 cache: packages
-before_install:
-  - cd fuse
 
 r_github_packages:
   - jimhester/covr
