### 2016-09-05 ab950f0
Added code of conduct
