+++ README.Rmd
@@ -20,7 +20,7 @@ knitr::opts_chunk$set(
 <!-- Edit the README.Rmd only!!! The README.md is generated automatically from README.Rmd. -->
 
 fuse: Framework for Understanding Structural Errors (Hydrological Modelling)
-======
+---------------
 
 [![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.14005.svg)](http://dx.doi.org/10.5281/zenodo.14005)
 [![Build Status](https://travis-ci.org/cvitolo/fuse.svg)](https://travis-ci.org/cvitolo/fuse.svg?branch=master)
@@ -28,6 +28,9 @@ fuse: Framework for Understanding Structural Errors (Hydrological Modelling)
 [![codecov.io](https://codecov.io/github/cvitolo/fuse/coverage.svg?branch=master)](https://codecov.io/github/cvitolo/fuse?branch=master)
 
 Implementation of the framework for hydrological modelling FUSE described in Clark et al. (2008) and based on the Fortran code provided by M. Clark in 2011. The package consists of two modules: Soil Moisture Accounting module (fusesma.sim) and Gamma routing module (fuserouting.sim). It also contains default parameter ranges (fusesma.ranges and fuserouting.ranges) and three data objects: DATA (sample input dataset), parameters (sample parameters) and modlist (list of FUSE model structures).
+
+
+Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
 
 ## Dependencies & installation instructions
 
