### 2016-09-05 #52 @cvitolo
@arfon I cloned the repository to [fuse](https://github.com/cvitolo/fuse) and moved the package to the root dir. That fixed all the problems with appveyor. Would it be possible to continue the review on the cloned repo?
