### 2016-09-05 #52 @arfon
> @arfon I cloned the repository to fuse and moved the package to the root dir. That fixed all the problems with appveyor. Would it be possible to continue the review on the cloned repo?

@cvitolo - just to clarify. You would like to replace the repository associated with this submission from https://github.com/cvitolo/r_fuse to be https://github.com/cvitolo/fuse?
