+++ README.Rmd
@@ -1,9 +1,7 @@
 ---
 output:
-  html_document: 
-    keep_md: yes
-    variant: markdown_github
-    toc: true
+  github_document: 
+  toc: true
 ---
 
 ```{r echo=FALSE}
@@ -23,6 +21,7 @@ fuse: Framework for Understanding Structural Errors (Hydrological Modelling)
 ---------------
 
 [![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.14005.svg)](http://dx.doi.org/10.5281/zenodo.14005)
+[![status](http://joss.theoj.org/papers/392a55daada04a86f95eaa8da134a28d/status.svg)](http://joss.theoj.org/papers/392a55daada04a86f95eaa8da134a28d)
 [![Build Status](https://travis-ci.org/cvitolo/fuse.svg)](https://travis-ci.org/cvitolo/fuse.svg?branch=master)
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/cvitolo/fuse?branch=master&svg=true)](https://ci.appveyor.com/project/cvitolo/fuse)
 [![codecov.io](https://codecov.io/github/cvitolo/fuse/coverage.svg?branch=master)](https://codecov.io/github/cvitolo/fuse?branch=master)
