+++ README.md
@@ -1,28 +1,22 @@
 
-
-
 <!-- Edit the README.Rmd only!!! The README.md is generated automatically from README.Rmd. -->
-
 fuse: Framework for Understanding Structural Errors (Hydrological Modelling)
----------------
+----------------------------------------------------------------------------
 
-[![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.14005.svg)](http://dx.doi.org/10.5281/zenodo.14005)
-[![Build Status](https://travis-ci.org/cvitolo/fuse.svg)](https://travis-ci.org/cvitolo/fuse.svg?branch=master)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/cvitolo/fuse?branch=master&svg=true)](https://ci.appveyor.com/project/cvitolo/fuse)
-[![codecov.io](https://codecov.io/github/cvitolo/fuse/coverage.svg?branch=master)](https://codecov.io/github/cvitolo/fuse?branch=master)
+[![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.14005.svg)](http://dx.doi.org/10.5281/zenodo.14005) [![status](http://joss.theoj.org/papers/392a55daada04a86f95eaa8da134a28d/status.svg)](http://joss.theoj.org/papers/392a55daada04a86f95eaa8da134a28d) [![Build Status](https://travis-ci.org/cvitolo/fuse.svg)](https://travis-ci.org/cvitolo/fuse.svg?branch=master) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/cvitolo/fuse?branch=master&svg=true)](https://ci.appveyor.com/project/cvitolo/fuse) [![codecov.io](https://codecov.io/github/cvitolo/fuse/coverage.svg?branch=master)](https://codecov.io/github/cvitolo/fuse?branch=master)
 
 Implementation of the framework for hydrological modelling FUSE described in Clark et al. (2008) and based on the Fortran code provided by M. Clark in 2011. The package consists of two modules: Soil Moisture Accounting module (fusesma.sim) and Gamma routing module (fuserouting.sim). It also contains default parameter ranges (fusesma.ranges and fuserouting.ranges) and three data objects: DATA (sample input dataset), parameters (sample parameters) and modlist (list of FUSE model structures).
 
-
 Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
 
-## Dependencies & installation instructions
+Dependencies & installation instructions
+----------------------------------------
 
 ### Dependencies
+
 The rdefra package depends on a number of CRAN packages. Check for missing dependencies and install them:
 
-
-```r
+``` r
 packs <- c('zoo', 'tgp', 'devtools')
 new.packages <- packs[!(packs %in% installed.packages()[,"Package"])]
 if(length(new.packages)) install.packages(new.packages)
@@ -32,29 +26,29 @@ if(length(new.packages)) install.packages(new.packages)
 
 You can install this package from Github with [devtools](https://github.com/hadley/devtools):
 
-
-```r
+``` r
 devtools::install_github("cvitolo/fuse")
 ```
 
 Load the rdefra package:
 
-
-```r
+``` r
 library(fuse)
 ```
 
-## Sample data and parameters
+Sample data and parameters
+--------------------------
+
 Load sample data (daily time step)
 
-```r
+``` r
 data(DATA)
 myDELTIM <- 1
 ```
 
[…]
