### 2016-09-06 #52 @arfon
> @arfon Yes,  please.  Would that be possible? Thanks!

Sure thing. The review is now associated with https://github.com/cvitolo/fuse. @pragyansmita the updated paper is [10.21105.joss.00052.pdf](https://github.com/openjournals/joss-reviews/files/456925/10.21105.joss.00052.pdf)
