### 2016-09-06 d5ee6ab
Fixed toc in README
