### 2016-09-06 87bdbb1
Updated v3.1 submitted to JOSS
