### 2016-09-06 5627238
Added lazydata to DESCRIPTION and updated paper with new doi
