+++ paper/paper.bib
@@ -1,9 +1,9 @@
 @Manual{fuse-archive,
   author       = {Claudia Vitolo and Peter Wells and Martin Dobias and Wouter Buytaert},
-  title        = {fuse: Framework for Understanding Structural Errors, R package version 3.0.0},
+  title        = {fuse: Framework for Understanding Structural Errors, R package version 3.1},
   year         = 2012,
-  doi          = {10.5281/zenodo.14005},
-  url          = {https://github.com/cvitolo/r_fuse}
+  doi          = {10.5281/zenodo.61639},
+  url          = {https://github.com/cvitolo/fuse}
 }
 
 @Manual{R-base,
