+++ NAMESPACE
@@ -9,5 +9,6 @@ export(readmd)
 export(readmd2var)
 importFrom(stats,runif)
 importFrom(tgp,lhs)
+importFrom(utils,data)
 importFrom(zoo,coredata)
 useDynLib(fuse)
