### 2016-09-07 a4ae292
Moved text about code of conduct to META
