+++ README.md
@@ -1,4 +1,15 @@
 
+-   [fuse: Framework for Understanding Structural Errors (Hydrological Modelling)](#fuse-framework-for-understanding-structural-errors-hydrological-modelling)
+-   [Dependencies & installation instructions](#dependencies-installation-instructions)
+    -   [Dependencies](#dependencies)
+    -   [Installation](#installation)
+-   [Sample data and parameters](#sample-data-and-parameters)
+-   [Applications](#applications)
+    -   [Example usage with 1 model structure](#example-usage-with-1-model-structure)
+    -   [Ensemble modelling](#ensemble-modelling)
+-   [Use fuse with hydromad](#use-fuse-with-hydromad)
+-   [Meta](#meta)
+
 <!-- Edit the README.Rmd only!!! The README.md is generated automatically from README.Rmd. -->
 fuse: Framework for Understanding Structural Errors (Hydrological Modelling)
 ----------------------------------------------------------------------------
@@ -6,8 +17,6 @@ fuse: Framework for Understanding Structural Errors (Hydrological Modelling)
 [![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.14005.svg)](http://dx.doi.org/10.5281/zenodo.14005) [![status](http://joss.theoj.org/papers/392a55daada04a86f95eaa8da134a28d/status.svg)](http://joss.theoj.org/papers/392a55daada04a86f95eaa8da134a28d) [![Build Status](https://travis-ci.org/cvitolo/fuse.svg)](https://travis-ci.org/cvitolo/fuse.svg?branch=master) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/cvitolo/fuse?branch=master&svg=true)](https://ci.appveyor.com/project/cvitolo/fuse) [![codecov.io](https://codecov.io/github/cvitolo/fuse/coverage.svg?branch=master)](https://codecov.io/github/cvitolo/fuse?branch=master)
 
 Implementation of the framework for hydrological modelling FUSE described in Clark et al. (2008) and based on the Fortran code provided by M. Clark in 2011. The package consists of two modules: Soil Moisture Accounting module (fusesma.sim) and Gamma routing module (fuserouting.sim). It also contains default parameter ranges (fusesma.ranges and fuserouting.ranges) and three data objects: DATA (sample input dataset), parameters (sample parameters) and modlist (list of FUSE model structures).
-
-Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
 
 Dependencies & installation instructions
 ----------------------------------------
@@ -237,6 +246,7 @@ summary(modfit)
 Meta
 ----
 
+-   Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
 -   Please [report any issues or bugs](https://github.com/cvitolo/fuse/issues).
 -   License: [GPL-3](https://opensource.org/licenses/GPL-3.0)
 -   Get citation information for the `fuse` package in R doing `citation(package = 'fuse')`
