### 2016-09-07 28b2c68
Check various r versions and rtools
