### 2016-09-07 #52 @cvitolo
@pragyansmita I don't know what prevents the fuse package to install on your system. The package seems to builds fine for [linux](https://travis-ci.org/cvitolo/fuse/builds/158181099) and [windows](https://ci.appveyor.com/project/cvitolo/fuse) systems. I have also tested various r versions and get no errors, as you can see [here](https://ci.appveyor.com/project/cvitolo/fuse). Do you mind to give it another try? Please let me know if you still have problems.

@arfon Many thanks for changing the repo link!
