### 2016-09-09 6595061
Re-arranged badges
