### 2016-09-09 08b5303
typos in modlist.rd
