### 2016-09-09 4c33ef8
Made fuse::fuseInfo and amca::PrepareTable consistent
