+++ R/fuseInfo.R
@@ -25,107 +25,195 @@ fuseInfo <- function(mid) {
                          "qbrate_2b","sareamax","axv_bexp","loglamb",
                          "tishape","timedelay")
 
-  # (1) rainfall errors
+  # (1) rainfall errors ########################################################
   if(model$rferr == 11) parameters$rferr_add <- TRUE                # additive_e
   if(model$rferr == 12) parameters$rferr_mlt <- TRUE                # multiplc_e
 
-  # (2) upper-layer architecture
+  # (2) upper-layer architecture ###############################################
   # Common parameters:
-  parameters$maxwatr_1 <- TRUE # maximum total storage in layer1 (mm)
-  parameters$fracten   <- TRUE # frac total storage as tension storage (-)
+  parameters$maxwatr_1 <- TRUE            # maximum total storage in layer1 (mm)
+  parameters$fracten   <- TRUE       # frac total storage as tension storage (-)
   # even with a single state, tension and free storage should be defined!
+
   # Model-specific parameters:
-  # if(model$arch1 == 21) {}            # onestate_1, no other parameters needed
-  # if(model$arch1 == 22) {}            # tension1_1, no other parameters needed
+
+  # onestate_1, no other parameters needed
+  # if(model$arch1 == 21) {}
+
+  # tension1_1, no other parameters needed
+  # if(model$arch1 == 22) {}
+
+  # tension2_1 = tension storage sub-divided into recharge and excess
   if(model$arch1 == 23) {
-    # tension2_1 = tension storage sub-divided into recharge and excess
-    parameters$frchzne <- TRUE # PRMS: frac tension storage in recharge zone (-)
+
+    # PRMS: frac tension storage in recharge zone (-)
+    parameters$frchzne <- TRUE
+
   }
 
-  # (3) lower-layer architecture / baseflow
+  # (3) lower-layer architecture / baseflow ####################################
   # Common parameters:
-  parameters$maxwatr_2 <- TRUE # maximum total storage in layer2 (mm)
+  # maximum total storage in layer2 (mm)
+  parameters$maxwatr_2 <- TRUE
+  # The params below are always used to calculate qbsat,
+  # this also means that loglamb and tishape are always calculated!
+  # mean value of the log-transformed topographic index (m)
+  parameters$loglamb   <- TRUE
+  # shape parameter for the topo index Gamma distribution (-)
+  parameters$tishape   <- TRUE
+  # baseflow exponent (-)
+  parameters$qb_powr   <- TRUE
+
   # Model-specific parameters:
+
+  # fixedsiz_2
+  # power-law relation, no parameters needed for the topo index distribution
   if(model$arch2 == 31) {
-    # fixedsiz_2
-    # power-law relation, no parameters needed for the topo index distribution
-    parameters$baserte   <- TRUE  # baseflow rate (mm day-1)
-    parameters$qb_powr   <- TRUE  # baseflow exponent (-)
-  }
-  if(model$arch2 == 32) {
-    # tens2pll_2 = tension reservoir plus two parallel tanks
-    parameters$percfrac  <- TRUE  # fraction of percolation to tension storage (-)
-    parameters$fprimqb   <- TRUE  # SAC: fraction of baseflow in primary resvr (-)
-    parameters$qbrate_2a <- TRUE  # baseflow depletion rate for primary resvr (day-1)
-    parameters$qbrate_2b <- TRUE  # baseflow depletion rate for secondary resvr (day-1)
-  }
-  if(model$arch2 == 33) {
[…]
