### 2016-10-06 #52 @pragyansmita
@cvitolo I am getting another error on following the install steps (Windows R 3.3.1 and Rtools 3.4)

> packs <- c('zoo', 'tgp', 'devtools')
> new.packages <- packs[!(packs %in% installed.packages()[,"Package"])]
> if(length(new.packages)) install.packages(new.packages)
> devtools::install_github("cvitolo/fuse")
Downloading GitHub repo cvitolo/fuse@master
from URL https://api.github.com/repos/cvitolo/fuse/zipball/master
Installing fuse
"C:/R/R-33~1.1/bin/x64/R" --no-site-file --no-environ --no-save --no-restore --quiet CMD  \
  INSTALL "C:/Users/ITS  \
  Admin/AppData/Local/Temp/Rtmp44Yu27/devtools5b06f71667c/cvitolo-fuse-4c33ef8"  \
  --library="C:/R/R-3.3.1/library" --install-tests

* installing *source* package 'fuse' ...
** libs

*** arch - i386
C:\Rtools\mingw_32\bin\nm.exe: fuse.o: File format not recognized
C:\Rtools\mingw_32\bin\nm.exe: modelStructure.o: File format not recognized
c:/Rtools/mingw_32/bin/g++ -shared -s -static-libgcc -o fuse.dll tmp.def fuse.o modelStructure.o -Ld:/Compiler/gcc-4.9.3/local330/lib/i386 -Ld:/Compiler/gcc-4.9.3/local330/lib -LC:/R/R-33~1.1/bin/i386 -lR
fuse.o: file not recognized: File format not recognized
collect2.exe: error: ld returned 1 exit status
no DLL was created
ERROR: compilation failed for package 'fuse'
* removing 'C:/R/R-3.3.1/library/fuse'
Error: Command failed (1)
> library(fuse)
Error in library(fuse) : there is no package called ‘fuse’

Is something missing in the steps to install?
