+++ NAMESPACE
@@ -5,8 +5,10 @@ export(fuseInfo)
 export(fuserouting.sim)
 export(fusesma.sim)
 export(generateParameters)
+export(model_param_table)
 export(readmd)
 export(readmd2var)
+importFrom(dplyr,bind_rows)
 importFrom(stats,runif)
 importFrom(tgp,lhs)
 importFrom(utils,data)
