### 2016-10-15 d06ebdd
Updated dependences (including R version)
