+++ README.Rmd
@@ -29,20 +29,14 @@ fuse: Framework for Understanding Structural Errors (Hydrological Modelling)
 
 Implementation of the framework for hydrological modelling FUSE described in Clark et al. (2008) and based on the Fortran code provided by M. Clark in 2011. The package consists of two modules: Soil Moisture Accounting module (fusesma.sim) and Gamma routing module (fuserouting.sim). It also contains default parameter ranges (fusesma.ranges and fuserouting.ranges) and three data objects: DATA (sample input dataset), parameters (sample parameters) and modlist (list of FUSE model structures).
 
-## Dependencies & installation instructions
-
-### Dependencies
-The rdefra package depends on a number of CRAN packages. Check for missing dependencies and install them:
+## Installation
+If you have already installed [R](https://cran.r-project.org/) (and [Rtools](https://cran.r-project.org/bin/windows/Rtools/) for Windows users), you can install this package's dependecies from CRAN:
 
 ```{r, eval=FALSE}
-packs <- c("DBI", "assertthat", "magrittr", "tibble", "maptree")
-new.packages <- packs[!(packs %in% installed.packages()[,"Package"])]
-if(length(new.packages)) install.packages(new.packages)
+install.packages(c("DBI", "assertthat", "magrittr", "tibble", "maptree"))
 ```
 
-### Installation
-
-You can install this package from Github with [devtools](https://github.com/hadley/devtools):
+and the package itself from Github with [devtools](https://github.com/hadley/devtools):
 
 ```{r, eval=FALSE}
 devtools::install_github("cvitolo/fuse")
