+++ DESCRIPTION
@@ -15,7 +15,7 @@ Description: Implementation of the framework for hydrological modelling FUSE,
     Gamma routing module (\code{routgamma.sim}). It also contains default parameter
     ranges (see \code{fuse.ranges} and \code{routgamma.ranges}) and three data
     objects: DATA, parameters and modlist.
-Depends: R (>= 3.1.0)
+Depends: R (>= 3.3.1)
 Imports: dplyr, zoo, tgp, stats
 LinkingTo: BH, Rcpp
 Suggests: testthat
