### 2016-10-15 25de6d6
Updated appveyor with Rtools34
