+++ appveyor.yml
@@ -21,8 +21,7 @@ environment:
   matrix:
   - R_VERSION: stable
   - R_VERSION: patched
-  - R_VERSION: oldrel
-    RTOOLS_VERSION: 33
+    RTOOLS_VERSION: 34
 
 build_script:
   - travis-tool.sh install_deps
