### 2016-10-15 3d7a77f
Added fusesma.ranges and fuserouting.ranges to NAMESPACE
