+++ README.Rmd
@@ -30,10 +30,10 @@ fuse: Framework for Understanding Structural Errors (Hydrological Modelling)
 Implementation of the framework for hydrological modelling FUSE described in Clark et al. (2008) and based on the Fortran code provided by M. Clark in 2011. The package consists of two modules: Soil Moisture Accounting module (fusesma.sim) and Gamma routing module (fuserouting.sim). It also contains default parameter ranges (fusesma.ranges and fuserouting.ranges) and three data objects: DATA (sample input dataset), parameters (sample parameters) and modlist (list of FUSE model structures).
 
 ## Installation
-If you have already installed [R](https://cran.r-project.org/) (and [Rtools](https://cran.r-project.org/bin/windows/Rtools/) for Windows users), you can install the dependecies from CRAN:
+If you have already installed [R](https://cran.r-project.org/) (and [Rtools](https://cran.r-project.org/bin/windows/Rtools/) for Windows users), you can install the dependencies for the examples below from CRAN:
 
 ```{r, eval=FALSE}
-install.packages(c("DBI", "assertthat", "magrittr", "tibble", "maptree"))
+install.packages(c("DBI", "assertthat", "magrittr", "tibble", "maptree", "tgp", "zoo", "qualV"))
 ```
 
 and the package itself from Github with [devtools](https://github.com/hadley/devtools):
@@ -65,7 +65,7 @@ names(DefaultRanges) <- c("Min","Max")
 Sample parameter set using Latin Hypercube method
 ```{r}
 numberOfRuns <- 100
-parameters <- lhs( numberOfRuns, as.matrix(DefaultRanges) )
+parameters <- tgp::lhs( numberOfRuns, as.matrix(DefaultRanges) )
 parameters <- data.frame(parameters)
 names(parameters) <- row.names(DefaultRanges)
 ```
@@ -96,7 +96,7 @@ plot(DATA$Q,type="l",xlab="",ylab="Streamflow [mm/day]")
 allQ <- data.frame(matrix(NA,ncol=numberOfRuns,nrow=dim(DATA)[1]))
 for (i in 1:numberOfRuns){
   allQ[,i] <- fuse(DATA, myMID, myDELTIM, parameters[i,])
-  lines(zoo(allQ[,i],order.by=index(DATA)),col="gray",lwd=0.1)
+  lines(zoo::zoo(allQ[,i],order.by=zoo::index(DATA)),col="gray",lwd=0.1)
 }
 lines(DATA$Q,col="black")
 ```
@@ -109,7 +109,6 @@ mids <- c(60, 230, 342, 426)
  
 Run a multi-model calibration using the Nash-Sutcliffe efficiency as objective function
 ```{r}
-library(qualV)
 indices <- rep(NA,4*numberOfRuns)
 discharges <- matrix(NA,ncol=4*numberOfRuns,nrow=dim(DATA)[1])
 kCounter <- 0
@@ -125,7 +124,7 @@ for (m in 1:4){
     
     Qrout <- fuse(DATA, myMID, myDELTIM, parameters[pid,])
  
-    indices[kCounter] <- EF(DATA$Q,Qrout)  
+    indices[kCounter] <- qualV::EF(DATA$Q,Qrout)  
     discharges[,kCounter] <- Qrout
     
     }
@@ -145,13 +144,13 @@ bestModel <- function(runNumber){
 }
 bestModel(bestRun)
  
-plot(coredata(DATA$Q),type="l",xlab="",ylab="Streamflow [mm/day]", lwd=0.5)
+plot(zoo::coredata(DATA$Q),type="l",xlab="",ylab="Streamflow [mm/day]", lwd=0.5)
  
[…]
