+++ man/fusesma.ranges.Rd
@@ -1,16 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/fusesma.R
 \name{fusesma.ranges}
 \alias{fusesma.ranges}
 \title{Function to define the parameter ranges for FUSE Soil Moisture Accounting module}
 \usage{
 fusesma.ranges()
 }
+\value{
+The function returns the list of parameter ranges for the FUSE Soil Moisture Accounting model.
+}
 \description{
-Parameter ranges for FUSE Soil Moisture Accounting module, as suggested in Clark et al. (2008)
+Parameter ranges for FUSE Soil Moisture Accounting module, as suggested in Clark et al. (2011).
 }
-\details{
-Function to define the parameter ranges for FUSE Soil Moisture Accounting module
+\examples{
+\dontrun{
+fusesma.ranges()
 }
-\author{
-Claudia Vitolo
+
 }
 
