+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: fuse
 Title: Framework for Understanding Structural Errors
-Version: 3.1
+Version: 3.2
 Authors@R: c(person("Claudia", "Vitolo", role = c("aut", "cre"),
                     email = "cvitolodev@gmail.com"),
 	           person("Peter", "Wells", role = "ctr"),

[…]
