### 2016-10-15 #52 @cvitolo
Hi @pragyansmita. I have updated the installation instructions in the README. Could you please try again? Many thanks!
