### 2016-12-17 cvitolo/fuse#1 @maelle
Clark/Fortran questions
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

I read in the README (& paper) "Implementation of the framework for hydrological modelling FUSE described in Clark et al. (2008) and based on the Fortran code provided by M. Clark in 2011. "

I'm absolutely not a licensing expert, so I just wonder whether the original Fortran code was released (and if so under which license -- but I guess it wasn't released since you mention it was kindly provided?), whether the author agreed to its use for a R package with the license you've chosen? Or maybe it doesn't play a role since your code is not a binding but rather a translation (and even improvement?) of the original code? Had you considered adding him as a contributor of the package? But I see DESCRIPTION too mention the original paper so this issue is probably useless. :-)
