### 2016-12-17 cvitolo/fuse#2 @maelle
Installation instructions
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

The README states `install.packages(c("DBI", "assertthat", "magrittr", "tibble", "maptree", "tgp", "zoo", "qualV"))`

But from DESCRIPTION the dependencies are
Imports: dplyr, zoo, tgp, stats
LinkingTo: BH, Rcpp
Suggests: testthat

Should the installation instructions be updated?

That said, I had no issue installing the package and am now running R CHECK 🎉
