### 2016-12-17 cvitolo/fuse#3 @maelle
The plots don't appear in the README
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

A cache issue?
