### 2016-12-17 cvitolo/fuse#2 @maelle
The installation instructions are in the vignette too, and also have a different list of dependencies.
