### 2016-12-17 cvitolo/fuse#4 @maelle
Name and format of DATA
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

The most important remark of this issue: the doc states DATA is a data.frame but it's actually a `zoo` object, could you update the doc or the object itself?

These are details and only suggestions:

* The name "DATA" is not very specific, why not name the sample data fuse_hydrological_data or something like that? If it's a time series, maybe even fuse_hydrological_timeseries?

* The names of the columns/variables could be replaced by the full name of the variables, e.g. "precipitation" instead of P.
