### 2016-12-17 cvitolo/fuse#4 @maelle
By the way since it is a `zoo` object, in the doc you could add "plot(DATA)" in Examples, it's quite nice one gets to see the time series so easily. :-)
