### 2016-12-17 cvitolo/fuse#1 @cvitolo
You made very good points here! The answers to your questions are certainly not obvious.

In 2011, the author of fuse (Fortran) just provided the code to interested researchers and groups. I contacted Martyn and asked him if I could add his name as contributor to my re-implementation of fuse in R (JOSS paper) and he said that was not necessary. But he put me in contact with his postdoc who has recently taken over the fuse (fortran) code and released it under a GPL-3 license here: https://github.com/naddor/fuse. Our intention, in the future, is to merge our efforts and make a unique code with Rinterface but we haven't worked out the details yet. Any suggestion is welcome!
