### 2016-12-17 05b5fa5
Made list of dependencies consistent across DESCRIPTION, README and vignette
