+++ vignettes/fuse_vignette.Rmd
@@ -23,7 +23,8 @@ Implementation of the framework for hydrological modelling FUSE described in Cla
 The rdefra package depends on a number of CRAN packages. Check for missing dependencies and install them:
 
 ```{r, eval=FALSE}
-packs <- c('zoo', 'tgp', 'devtools')
+packs <- c("dplyr", "zoo", "tgp", "stats", "utils", "BH", "Rcpp", "testthat",
+           "devtools")
 new.packages <- packs[!(packs %in% installed.packages()[,"Package"])]
 if(length(new.packages)) install.packages(new.packages)
 ```
