### 2016-12-17 cvitolo/fuse#2 @cvitolo
Thanks for spotting this. The list of dependencies are now consistent across README, DESCRIPTION and vignette.
