### 2016-12-17 d49fb7b
Removed vignette cache
