+++ vignettes/fuse_vignette.Rmd
@@ -12,7 +12,7 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   warning = FALSE,
   message = FALSE,
-  cache = TRUE
+  cache = FALSE
 )
 ```
 
