### 2016-12-17 cvitolo/fuse#5 @maelle
Name and format of modlist and modliststring
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

These are only suggestions and questions:

* I'd be in favour of naming `modlist` more explicitely, e.g. `fuse_models`, and maybe even without "list" in the name since it's a `data.frame` (I understand the reason for calling it list, but I guess that because list is an object class in R it might be confusing?).

* I do not understand the use of `modliststring`. When looking for its occurrences in the repo I could only find the documentation. If the only role of `modliststring` is to explain the codes, why are abbreviations used? Maybe it'd make sense to no longer have `modliststring` and instead have `modlist` and a data.frame per type of code with a column "code" and a column "meaning", or to have only `modlist` and its documentation. Or to have only `modliststring` and the documentation, i.e. use the abbreviations instead of the codes, and explain the abbrevations in the documentation?

* The names of the variables themselves could be a bit more explicit, e.g. `esoil` could be `soil_evaporation`. Unless those are all abbreviations commonly used in hydrology?
