### 2016-12-17 cvitolo/fuse#5 @maelle
"On Clark's model list, the parent models (Topmodel, VIC, PRMS and Sacramento) have id = 60, 230, 342 and 426 respectively." -> why not add a variable `clark_name`?
