### 2016-12-17 cvitolo/fuse#6 @maelle
Related packages in the environmetrics CRAN Task View?
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

* Could you add a "similar packages" section to the README and vignette to mention the CRAN Task View environmetrics https://cran.r-project.org/web/views/Environmetrics.html and whether some packages do something similar or whether some packages have an output that could be used with `fuse`? This would help see how unique `fuse` is and whether there are alternative/complementary packages in R.

* When you submit the package to CRAN, I'd encourage you to contact the Task View maintainer to get `fuse` added to it but you might have already thought of this anyway.
