### 2016-12-17 cvitolo/fuse#5 @maelle
I see you use the codes (and not the abbreviations) internally, e.g. in

```r
  model <- data.frame(matrix(readmd2var(mid,number=TRUE),ncol=8,nrow=1))
names(model) <- names(readmd2var(mid,number=TRUE))
```

However for the user codes are not really useful, are they?

I think it's great to have a function for showing values for a single model ID only, but why have two of them, i.e. `readmd2var` and `readmd`? You could export only one with focus on "pretty" (after all otherwise the user can filter from `modlist` or `modliststring`) and not export the other one which you use internally?
