### 2016-12-17 cvitolo/fuse#7 @maelle
Functions naming
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

This is only a suggestion: many exported functions of the package have a name starting with "fuse", why not all of them?
