### 2016-12-17 cvitolo/fuse#5 @maelle
Oh I see there's `fuseInfo` too. It doesn't seem easy to know which one to use, the three `fuseInfo`, `readmd` and `readmd2var` do something really similar, right?
