### 2016-12-17 cvitolo/fuse#8 @maelle
Name of and names in parameters
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

This is again only a suggestion and quite similar to the suggestions made for the other datasets so I should maybe have grouped them, sorry.

* parameters could have a more specific name, e.g. fuse_parameters.

* the column names could be even more explicit (but I don't know about usual hydrology abbreviation so maybe I'm wrong), e.g. `tishape` could be `topology_index_shape`
