### 2016-12-17 cvitolo/fuse#9 @maelle
fuserouting.ranges and fusesma.ranges
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

The title of the doc of `fuserouting.ranges` is "Function to define the parameter ranges for GAMMA routing module" but the function has no parameter, it only returns the default parameters so it's not for defining something.

* I'd propose to change the title to "default parameter ranges" or something like that.

* maybe the name of the function could even contain the word "default"?
