### 2016-12-17 cvitolo/fuse#10 @maelle
fuserouting.sim
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

The doc of fuserouting.sim should mention the class of the arguments, e.g. whether `deltim` is a numeric.
