### 2016-12-17 cvitolo/fuse#9 @maelle
I have the same comments for `fusesma.ranges`.
