### 2016-12-17 cvitolo/fuse#8 @maelle
I guess if you change the names of the variables it'd mean also changing the name of the parameters of `fusesma.sim`, not fun but maybe user-friendlier?
