### 2016-12-17 cvitolo/fuse#11 @maelle
fusesma.sim/fuserouting.sim/fuse
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

Why not have the random seed as argument of `fusesma.sim` and `fuserouting.sim` and `fuse` (or of `fuse`only since it calls both other functions), with a default value, to make results reproducible even if the user forgets about the seed?
