### 2016-12-17 cvitolo/fuse#12 @maelle
fuse
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

I wonder if the name of the function `fuse` is too confusing. Maybe you could rename it to something with "simulation" or "sim" in the name?
