### 2016-12-17 cvitolo/fuse#2 @maelle
Awesome!
