### 2016-12-17 cvitolo/fuse#1 @maelle
At least out a link to the Fortran implementation then?

If Fortran isn't used a lot, maybe this postdoc can come help you with the boring issues I've opened 😀
