### 2016-12-17 cvitolo/fuse#12 @maelle
By the way are the two modules composing the functions used on their own a lot? If not, I'd only export the "meta-function".
