### 2016-12-17 cvitolo/fuse#3 @cvitolo
Thanks for spotting this! I had set the code chunck that generates the plot with eval = FALSE... which did not evaluate the code and therefore did not plot the figure. This is now sorted!

I also simplified the readme (according to ggplot2 [here](https://github.com/tidyverse/ggplot2)) and left all the examples in the vignette. This reduces redundancies considerably.
