### 2016-12-17 cvitolo/fuse#3 @maelle
@cvitolo great! Maybe you could generate a website for the package with `pkgdown` to make it easier to browse the vignette then?
