### 2016-12-17 cvitolo/fuse#3 @maelle
I don't know if you've used `pkgdown` before, I use it [here](https://github.com/ropensci/ropenaq) for instance.
