### 2016-12-17 cvitolo/fuse#1 @cvitolo
The github link above relates to the new version of fuse, not the one I originally worked on.
Do you think that would confuse users?
