### 2016-12-17 cvitolo/fuse#6 @maelle
I see in the vignette that you discuss `hydromad`, maybe you could mention other things that one can do with `hydromad`?
