### 2016-12-17 cvitolo/fuse#3 @cvitolo
never used `pkgdown` before, I'm going to have a look now.
