### 2016-12-17 cvitolo/fuse#1 @maelle
Ah yes it might... maybe at the bottom of the README it could make sense to say something like "For Fortran users, please refer to the new version of fuse"?

What differences are there between the old and the new version? It could make sense to mention that you plan to incorporate all new features of Fortran fuse in your R fuse and that the development of Fortran fuse might even stop in favor of R fuse?
