### 2016-12-17 a512b4f
Simplified README, all the examples are in the vignette
