+++ vignettes/fuse_vignette.Rmd
@@ -12,7 +12,8 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   warning = FALSE,
   message = FALSE,
-  cache = FALSE
+  cache = FALSE,
+  eval = TRUE
 )
 ```
 
@@ -20,28 +21,24 @@ knitr::opts_chunk$set(
 Implementation of the framework for hydrological modelling FUSE described in Clark et al. (2008) and based on the Fortran code provided by M. Clark in 2011. The package consists of two modules: Soil Moisture Accounting module (fusesma.sim) and Gamma routing module (fuserouting.sim). It also contains default parameter ranges (fusesma.ranges and fuserouting.ranges) and three data objects: DATA (sample input dataset), parameters (sample parameters) and modlist (list of FUSE model structures).
 
 ### Dependencies
-The rdefra package depends on a number of CRAN packages. Check for missing dependencies and install them:
+The fuse package, as well as the examples in this vignette, depend on a number of CRAN packages. Check for missing dependencies and install them:
 
-```{r, eval=FALSE}
+```{r, eval = FALSE}
 packs <- c("dplyr", "zoo", "tgp", "stats", "utils", "BH", "Rcpp", "testthat",
-           "devtools")
-new.packages <- packs[!(packs %in% installed.packages()[,"Package"])]
+           "qualV", "devtools")
+new.packages <- packs[!(packs %in% installed.packages()[, "Package"])]
 if(length(new.packages)) install.packages(new.packages)
 ```
 
 ### Installation
-
 You can install this package from Github with [devtools](https://github.com/hadley/devtools):
-
-```{r, eval=FALSE}
-library(devtools)
-install_github("cvitolo/r_fuse", subdir = "fuse")
+```{r, eval = FALSE}
+devtools::install_github("cvitolo/fuse")
 ```
 
-Load the rdefra package:
-
+Load the fuse package:
 ```{r}
-library(fuse)
+library("fuse")
 ```
 
 ### Data and parameters
@@ -52,14 +49,14 @@ myDELTIM <- 1
 ```
 
 Define parameter ranges
-```{r}
+```{r, eval = FALSE}
 DefaultRanges <- data.frame(t(data.frame(c(fusesma.ranges(),
                                            fuserouting.ranges()))))
-names(DefaultRanges) <- c("Min","Max")
+names(DefaultRanges) <- c("Min", "Max")
 ```
 
 Sample parameter set using Latin Hypercube method
-```{r}
+```{r, eval = FALSE}
 numberOfRuns <- 100
 parameters <- lhs( numberOfRuns, as.matrix(DefaultRanges) )
 parameters <- data.frame(parameters)
@@ -68,6 +65,7 @@ names(parameters) <- row.names(DefaultRanges)
 
 Alternatively, sample parameter set using built-in function
 ```{r}
+set.seed(1)
 parameters <- generateParameters(100)
 ```
 
@@ -79,23 +77,29 @@ myMID <- 60
 
 Use the built-in function to run FUSE for the 1st sampled parameter set
 ```{r}
-x <- fuse(DATA, myMID, myDELTIM, parameters[1,])
+streamflow <- fuse(DATA, myMID, myDELTIM, parameters[1,])
 
[…]
