+++ README.Rmd
@@ -1,7 +1,6 @@
 ---
 output:
-  github_document: 
-    toc: true
+  github_document
 ---
 
 ```{r echo=FALSE}
@@ -10,8 +9,8 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   warning = FALSE,
   message = FALSE,
-  cache = TRUE,
-  eval = FALSE
+  cache = FALSE,
+  eval = TRUE
 )
 ```
 
@@ -29,203 +28,46 @@ fuse: Framework for Understanding Structural Errors (Hydrological Modelling)
 
 Implementation of the framework for hydrological modelling FUSE described in Clark et al. (2008) and based on the Fortran code provided by M. Clark in 2011. The package consists of two modules: Soil Moisture Accounting module (fusesma.sim) and Gamma routing module (fuserouting.sim). It also contains default parameter ranges (fusesma.ranges and fuserouting.ranges) and three data objects: DATA (sample input dataset), parameters (sample parameters) and modlist (list of FUSE model structures).
 
-## Installation
-If you have already installed [R](https://cran.r-project.org/) (and [Rtools](https://cran.r-project.org/bin/windows/Rtools/) for Windows users), you can install the dependencies for the examples below from CRAN:
+### Dependencies
+The fuse package, as well as the examples in the vignette, depend on a number of CRAN packages. Check for missing dependencies and install them:
 
 ```{r, eval=FALSE}
 packs <- c("dplyr", "zoo", "tgp", "stats", "utils", "BH", "Rcpp", "testthat",
-           "devtools")
-new.packages <- packs[!(packs %in% installed.packages()[,"Package"])]
+           "qualV", "devtools")
+new.packages <- packs[!(packs %in% installed.packages()[, "Package"])]
 if(length(new.packages)) install.packages(new.packages)
 ```
 
-and the package itself from Github with [devtools](https://github.com/hadley/devtools):
-
+## Installation
+You can install this package from Github with [devtools](https://github.com/hadley/devtools):
 ```{r, eval=FALSE}
 devtools::install_github("cvitolo/fuse")
 ```
 
 Load the package:
-
 ```{r}
-library(fuse)
+library("fuse")
 ```
 
-## Sample data and parameters
-Load sample data (daily time step)
-```{r}
+## Usage
+The simplest way to test `fuse` is to use the dummy data provided with this package, sample the parameter range using the built-in function `generateParameters()` and run the model with 1 model structure (e.g. TOPMODEL):
+```{r, eval = TRUE}
+# Load the dummy data
 data(DATA)
-myDELTIM <- 1
+
+# Sample the parameter range
+parameters <- generateParameters(1)
+
+# Run fuse 
+streamflow <- fuse(DATA = DATA, 
+                   mid = 60,       # Model structure = TOPMODEL
+                   deltim = 1,     # Daily time step
+                   ParameterSet = parameters)
+
+plot(streamflow, xlab = "", ylab = "Streamflow [mm/day]")
 ```
 
-Define parameter ranges
-```{r}
-DefaultRanges <- data.frame(t(data.frame(c(fusesma.ranges(),
-                                           fuserouting.ranges()))))
-names(DefaultRanges) <- c("Min","Max")
-```
-
-Sample parameter set using Latin Hypercube method
-```{r}
-numberOfRuns <- 100
[…]
