+++ README.md
@@ -1,13 +1,4 @@
 
--   [fuse: Framework for Understanding Structural Errors (Hydrological Modelling)](#fuse-framework-for-understanding-structural-errors-hydrological-modelling)
--   [Installation](#installation)
--   [Sample data and parameters](#sample-data-and-parameters)
--   [Applications](#applications)
-    -   [Example usage with 1 model structure](#example-usage-with-1-model-structure)
-    -   [Ensemble modelling](#ensemble-modelling)
--   [Use fuse with hydromad](#use-fuse-with-hydromad)
--   [Meta](#meta)
-
 <!-- Edit the README.Rmd only!!! The README.md is generated automatically from README.Rmd. -->
 fuse: Framework for Understanding Structural Errors (Hydrological Modelling)
 ----------------------------------------------------------------------------
@@ -18,19 +9,21 @@ fuse: Framework for Understanding Structural Errors (Hydrological Modelling)
 
 Implementation of the framework for hydrological modelling FUSE described in Clark et al. (2008) and based on the Fortran code provided by M. Clark in 2011. The package consists of two modules: Soil Moisture Accounting module (fusesma.sim) and Gamma routing module (fuserouting.sim). It also contains default parameter ranges (fusesma.ranges and fuserouting.ranges) and three data objects: DATA (sample input dataset), parameters (sample parameters) and modlist (list of FUSE model structures).
 
-Installation
-------------
+### Dependencies
 
-If you have already installed [R](https://cran.r-project.org/) (and [Rtools](https://cran.r-project.org/bin/windows/Rtools/) for Windows users), you can install the dependencies for the examples below from CRAN:
+The fuse package, as well as the examples in the vignette, depend on a number of CRAN packages. Check for missing dependencies and install them:
 
 ``` r
 packs <- c("dplyr", "zoo", "tgp", "stats", "utils", "BH", "Rcpp", "testthat",
-           "devtools")
-new.packages <- packs[!(packs %in% installed.packages()[,"Package"])]
+           "qualV", "devtools")
+new.packages <- packs[!(packs %in% installed.packages()[, "Package"])]
 if(length(new.packages)) install.packages(new.packages)
 ```
 
-and the package itself from Github with [devtools](https://github.com/hadley/devtools):
+Installation
+------------
+
+You can install this package from Github with [devtools](https://github.com/hadley/devtools):
 
 ``` r
 devtools::install_github("cvitolo/fuse")
@@ -39,201 +32,33 @@ devtools::install_github("cvitolo/fuse")
 Load the package:
 
 ``` r
-library(fuse)
+library("fuse")
 ```
 
-Sample data and parameters
---------------------------
+Usage
+-----
 
-Load sample data (daily time step)
+The simplest way to test `fuse` is to use the dummy data provided with this package, sample the parameter range using the built-in function `generateParameters()` and run the model with 1 model structure (e.g. TOPMODEL):
 
 ``` r
+# Load the dummy data
 data(DATA)
-myDELTIM <- 1
+
+# Sample the parameter range
+parameters <- generateParameters(1)
+
+# Run fuse 
+streamflow <- fuse(DATA = DATA, 
[…]
