### 2016-12-17 cvitolo/fuse#13 @maelle
test coverage
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

Looking at codecov.io, it seems that the test coverage could be improved with a few more tests. :-)
