### 2016-12-17 cvitolo/fuse#14 @maelle
was the package used in any published work?
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

Because in that case you might want to mention that as example somewhere in the documentation?
