### 2016-12-17 107e54d
README: added link to Nans Addor's github fuse repository
