+++ README.Rmd
@@ -51,7 +51,7 @@ library("fuse")
 
 ## Usage
 The simplest way to test `fuse` is to use the dummy data provided with this package, sample the parameter range using the built-in function `generateParameters()` and run the model with 1 model structure (e.g. TOPMODEL):
-```{r, eval = TRUE}
+```{r example, eval = TRUE}
 # Load the dummy data
 data(DATA)
 
@@ -67,7 +67,9 @@ streamflow <- fuse(DATA = DATA,
 plot(streamflow, xlab = "", ylab = "Streamflow [mm/day]")
 ```
 
-For details and more complex examples (e.g. multi-model ensemble), please refer to the [vignette](vignettes/fuse_vignette.Rmd).
+For details and more complex examples (e.g. multi-model ensemble), please refer to the [vignette](vignettes/fuse_vignette.Rmd). 
+
+Fortran users are advised to refer to the latest version of fuse available [here](https://github.com/naddor/fuse).
 
 ## Meta
 
