### 2016-12-17 cvitolo/fuse#1 @cvitolo
I have added a link to the latest fortran fuse in the readme but I don't remember exactly what the new features are so I did not mention that. Also, the plan is to keep the fortran code and adapt my package to become the r interface to the fortran code. But I have no idea when that is going to happen...I'll add future plans to the README when I have a better idea ;)
