### 2016-12-17 cvitolo/fuse#1 @cvitolo
@masalmon Are you ok if I close this issue?
