### 2016-12-17 #52 @masalmon
This is a cool package, congrats on making the initial Fortran code in a package and available to R users!

## Reviewer questions
### General checks

No, the development version is 3.2 and the release version is 3.1.

### Functionality

- [ x] **Installation:** Does installation proceed as outlined in the documentation?
- [ x] **Functionality:** Have the functional claims of the software been confirmed?
- [ x] **Performance:** Have any performance claims of the software been confirmed?
### Documentation
- [x ] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Yes. Actually the part "The fuse framework takes as input rainfall and potential evapotranspiration time series (areal averages over the river catchment area) and returns a simulated time series of river discharges. It can be used to understand the variability of expected hydrological responses based on model structures. " of the paper could even be at the beginning of the README & vignette (or in the help page for the package.) too.

- [x ] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
- [ x] **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I have opened quite a few issues with suggestions for the documentation, but they are nearly all only suggestions.
- [x ] **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified?

The test coverage could be increased a bit though.

- [x ] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

The only thing that's not stated at the bottom of the README is whether code contributions are welcome.
### Software paper

Paper PDF: [10.21105.joss.00052.pdf](https://github.com/openjournals/joss-reviews/files/456925/10.21105.joss.00052.pdf)
- [ x] **Authors:** Does the `paper.md` file include a list of authors with their affiliations?
- [ x] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
- [ x] **References:** Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?
