### 2016-12-17 cvitolo/fuse#1 @maelle
Yes! Even doing it myself 😀
