### 2016-12-17 cvitolo/fuse#15 @maelle
NEWS.md
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

I'd recommend adding a NEWS file to the package (`devtools::use_news_md()` will create the template) for keeping track of changes introduced in the different version.
