+++ README.Rmd
@@ -2,7 +2,6 @@
 output:
   github_document
 ---
-
 ```{r echo=FALSE}
 knitr::opts_chunk$set(
   comment = "#>",
@@ -13,11 +12,10 @@ knitr::opts_chunk$set(
   eval = TRUE
 )
 ```
+fuse: Framework for Understanding Structural Errors ([HTML website](https://cvitolo.github.io/fuse/))
+======
 
 <!-- Edit the README.Rmd only!!! The README.md is generated automatically from README.Rmd. -->
-
-fuse: Framework for Understanding Structural Errors ([HTML website](https://cvitolo.github.io/fuse/))
-====================================================================
 
 [![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.14005.svg)](http://dx.doi.org/10.5281/zenodo.14005)
 [![status](http://joss.theoj.org/papers/392a55daada04a86f95eaa8da134a28d/status.svg)](http://joss.theoj.org/papers/392a55daada04a86f95eaa8da134a28d)
