+++ README.md
@@ -1,8 +1,8 @@
 
-<!-- Edit the README.Rmd only!!! The README.md is generated automatically from README.Rmd. -->
 fuse: Framework for Understanding Structural Errors ([HTML website](https://cvitolo.github.io/fuse/))
 =====================================================================================================
 
+<!-- Edit the README.Rmd only!!! The README.md is generated automatically from README.Rmd. -->
 [![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.14005.svg)](http://dx.doi.org/10.5281/zenodo.14005) [![status](http://joss.theoj.org/papers/392a55daada04a86f95eaa8da134a28d/status.svg)](http://joss.theoj.org/papers/392a55daada04a86f95eaa8da134a28d)
 
 [![Build Status](https://travis-ci.org/cvitolo/fuse.svg)](https://travis-ci.org/cvitolo/fuse.svg?branch=master) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/cvitolo/fuse?branch=master&svg=true)](https://ci.appveyor.com/project/cvitolo/fuse) [![codecov.io](https://codecov.io/github/cvitolo/fuse/coverage.svg?branch=master)](https://codecov.io/github/cvitolo/fuse?branch=master)
