+++ README.Rmd
@@ -12,7 +12,7 @@ knitr::opts_chunk$set(
   eval = TRUE
 )
 ```
-fuse: Framework for Understanding Structural Errors ([HTML website](https://cvitolo.github.io/fuse/))
+Framework for Understanding Structural Errors
 ======
 
 <!-- Edit the README.Rmd only!!! The README.md is generated automatically from README.Rmd. -->
