+++ README.Rmd
@@ -15,6 +15,8 @@ knitr::opts_chunk$set(
 
 <!-- Edit the README.Rmd only!!! The README.md is generated automatically from README.Rmd. -->
 
+## fuse: an R package implementing the Framework for Understanding Structural Errors
+
 [![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.14005.svg)](http://dx.doi.org/10.5281/zenodo.14005)
 [![status](http://joss.theoj.org/papers/392a55daada04a86f95eaa8da134a28d/status.svg)](http://joss.theoj.org/papers/392a55daada04a86f95eaa8da134a28d)
 
@@ -34,7 +36,7 @@ new.packages <- packs[!(packs %in% installed.packages()[, "Package"])]
 if(length(new.packages)) install.packages(new.packages)
 ```
 
-## Installation
+### Installation
 You can install this package from Github with [devtools](https://github.com/hadley/devtools):
 ```{r, eval=FALSE}
 devtools::install_github("cvitolo/fuse")
@@ -45,7 +47,7 @@ Load the package:
 library("fuse")
 ```
 
-## Usage
+### Usage
 The simplest way to test `fuse` is to use the dummy data provided with this package, sample the parameter range using the built-in function `generateParameters()` and run the model with 1 model structure (e.g. TOPMODEL):
 ```{r example, eval = TRUE}
 # Load the dummy data
@@ -67,7 +69,7 @@ For details and more complex examples (e.g. multi-model ensemble), please refer 
 
 Fortran users are advised to refer to the latest version of fuse available [here](https://github.com/naddor/fuse).
 
-## Meta
+### Meta
 
 * Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
 * Please [report any issues or bugs](https://github.com/cvitolo/fuse/issues).
