### 2016-12-17 cvitolo/fuse#3 @cvitolo
I couldn't figure out how to generate a website using pkgdown from the vignette file, so I decided to opt for the easy option: generated the website using the settings->GitHub Pages.

Added a link to the title.
