### 2016-12-17 cvitolo/fuse#3 @maelle
As you wish, but if you have any `pkgdown` question don't hesitate to ask. For ropenaq I think I only had to install pkgdown and then run:

```r
pkgdown::init()
pkgdown::build_site()
```
