### 2016-12-17 cvitolo/fuse#3 @maelle
After that I had to change the settings to, to docs/, so that Github could use files in docs for the website. But clearly it doesn't matter to me how you produce your website, awesome that you now got one, maybe we can close this issue :-)
