### 2016-12-17 cvitolo/fuse#3 @cvitolo
pkgdown is great! Just tried to build the website a is surprisingly simple. Thanks a lot for the tip!
