### 2016-12-17 586a262
Renamed dummy dataset DATA to fuse_hydrological_timeseries and updated documentation
