### 2016-12-17 cvitolo/fuse#4 @cvitolo
I have changed the name DATA to fuse_hydrological_timeseries and updated the documentation (format is now zoo) and added plot(fuse_hydrological_timeseries) to the examples in the doc.

I cannot changed the name of the name of the columns because this is a convention in modelling frameworks (e.g. Fortran fuse and hydromad).
