### 2016-12-17 fa6b569
Added comments to readmd2var, readmd and fuseInfo
