### 2016-12-17 cvitolo/fuse#5 @cvitolo
The names I used in the package reflect the original names in the fortran code. Keeping consistent names for the various objects allow me to track the changes between the fortran and R code.
For this reason I cannot change the names of `modlist` or variables (e.g. esoil).

I need to keep modliststring because this is the list used in the fortran code. I had to convert the strings into numbers (modlist) because R handles strings very slowly. Keeping both modlist and modliststring allows me to look-up what string corresponds to what number. Hope this makes sense.

I cannot create a `clark_name` because there are names only for 4 models (which are considered parent models).

readmd2var, readmd and fuseInfo have different goals:
- readmd2var is used by the model itself to generate the model structure from the id number (mid)
- readmd is used for reporting, to generate tables (usually via xtable) to list model structures and describe each component.
- fuseInfo is used to generate tables to summarise which model component and parameters are used given the model id. This is important for calculating the posterior distribution of parameter values and model structures.
I have added these comment to the related documentation.
