### 2016-12-18 cvitolo/fuse#5 @maelle
Ok, this all makes sense, thanks!
