### 2016-12-18 cvitolo/fuse#4 @maelle
Ok, cool!
