### 2016-12-18 800c854
Added a section called 'Other packages and future developments' to the vignette.
