### 2016-12-18 cvitolo/fuse#6 @cvitolo
Great suggestion! I have added a section called 'Other packages and future developments' to the vignette. I have not mentioned that in the README to avoid redundancies.
