### 2016-12-18 9daec1e
Added more info on hydromad capabilities
