### 2016-12-18 cvitolo/fuse#6 @cvitolo
Also added more info on hydromad capabilities to the vignette.
