### 2016-12-18 cvitolo/fuse#7 @cvitolo
The functions that start with 'fuse' are part of the original modelling framework (translation from the Fortran code), those that do not start with fuse are extra utility functions I developed. I'd like to keep separate to remember which function is also available in the fortran code.
