### 2016-12-18 cvitolo/fuse#8 @cvitolo
The names of the parameters are the same as in the fortran code. I totally agree that they could be more explicit but, in the light of future developments using the latest fortran code, it is important I keep the same naming convention.
