### 2016-12-18 cvitolo/fuse#9 @cvitolo
The functions fuserouting.ranges and fusesma.ranges only exist to make fuse compatible with hydromad. It is a requirement of hydromad that the routing module is accompanied by a function defining the default parameter ranges (same for the sma module). The naming convention is also dictated by hydromad, therefore it cannot be changed...sorry!
