### 2016-12-18 cvitolo/fuse#10 @cvitolo
Thanks, this is now done.
