### 2016-12-18 19ccfd1
fuserouting.sim now mentions the class of the arguments (https://github.com/cvitolo/fuse/issues/10)
