### 2016-12-18 857b473
Added class to arguments in fusesma.sim
