### 2016-12-18 cvitolo/fuse#11 @cvitolo
The functions fusesma.sim, fuserouting.sim and fuse do not generate internally the parameter sets but uses pre-defined ones because there are various ways of sampling the parameter ranges. The function generateParameters() only uses one of them (latin hypercube). For this reason, the parameter sets should be generated prior to using fuse and not within it.
