### 2016-12-18 cvitolo/fuse#12 @cvitolo
The names of the functions are designed to be consistent with the naming convention adopted in hydromad and other hydrological models (e.g. topmodel).

fusesma.sim and fuserouting.sim can be used on their own, depending on the local catchment under study (e.g. area of the catchment, slope, etc).
