### 2016-12-18 22ff52f
Added references to previous work in JOSS paper
