+++ paper/paper.bib
@@ -24,3 +24,85 @@
   number = {W00B02},
   doi = {10.1029/2007WR006735},
 }
+
+@article{Vitolo2015185,
+  title = "Web technologies for environmental Big Data ",
+  journal = "Environmental Modelling & Software ",
+  volume = "63",
+  number = "",
+  pages = "185 - 198",
+  year = "2015",
+  note = "",
+  issn = "1364-8152",
+  doi = "http://dx.doi.org/10.1016/j.envsoft.2014.10.007",
+  url = "http://www.sciencedirect.com/science/article/pii/S1364815214002965",
+  author = "Claudia Vitolo and Yehia Elkhatib and Dominik Reusser and Christopher J.A. Macleod and Wouter Buytaert",
+  keywords = "Web-based modelling",
+  keywords = "Big Data",
+  keywords = "Web services",
+  keywords = "OGC standards "
+}
+
+@article{Wilkinson201538,
+title = "A cloud based tool for knowledge exchange on local scale flood risk ",
+journal = "Journal of Environmental Management ",
+volume = "161",
+number = "",
+pages = "38 - 50",
+year = "2015",
+note = "",
+issn = "0301-4797",
+doi = "http://dx.doi.org/10.1016/j.jenvman.2015.06.009",
+url = "http://www.sciencedirect.com/science/article/pii/S0301479715301031",
+author = "M.E. Wilkinson and E. Mackay and P.F. Quinn and M. Stutter and K.J. Beven and C.J.A. MacLeod and M.G. Macklin and Y. Elkhatib and B. Percy and C. Vitolo and P.M. Haygarth",
+keywords = "Cloud computing",
+keywords = "Flooding",
+keywords = "Stakeholder engagement",
+keywords = "Rural land management",
+keywords = "Local \{EVOp\} Flooding Tool (LEFT) "
+}
+
+@Manual{topmodel,
+    title = {topmodel: Implementation of the hydrological model TOPMODEL in R},
+    author = {Wouter Buytaert},
+    year = {2011},
+    note = {R package version 0.7.2-2},
+    url = {https://CRAN.R-project.org/package=topmodel},
+}
+
+@Manual{dynatopmodel,
+    title = {dynatopmodel: Implementation of the Dynamic TOPMODEL Hydrological Model},
+    author = {Peter Metcalfe and Keith Beven and Jim Freer},
+    year = {2016},
+    note = {R package version 1.1},
+    url = {https://CRAN.R-project.org/package=dynatopmodel},
+  }
+
+@Manual{wasim,
+    title = {wasim: Visualisation and analysis of output files of the hydrological
+model WASIM},
+    author = {Dominik Reusser and Till Francke},
+    year = {2011},
+    note = {R package version 1.1.2},
+    url = {https://CRAN.R-project.org/package=wasim},
+  }
+
+@article{Andrews20111171,
+title = "An open software environment for hydrological model assessment and development ",
+journal = "Environmental Modelling & Software ",
+volume = "26",
+number = "10",
+pages = "1171 - 1185",
+year = "2011",
+note = "",
+issn = "1364-8152",
+doi = "http://dx.doi.org/10.1016/j.envsoft.2011.04.006",
+url = "http://www.sciencedirect.com/science/article/pii/S1364815211001046",
+author = "F.T. Andrews and B.F.W. Croke and A.J. Jakeman",
+keywords = "Model evaluation",
+keywords = "Hydrological models",
+keywords = "Modelling frameworks",
+keywords = "Unit hydrograph",
+keywords = "Event separation",
+keywords = "R "
+}
