### 2016-12-18 cvitolo/fuse#14 @cvitolo
Great idea thanks! I have added this to the JOSS paper.
