### 2016-12-18 f3da595
Added NEWS.md
