+++ NEWS.md
@@ -1,0 +1,4 @@
+# fuse 3.2
+
+* Added a `NEWS.md` file to track changes to the package.
+* Version reviewed for publication in the JOSS.
