### 2016-12-18 cvitolo/fuse#15 @cvitolo
Thanks, this is now done.
