### 2016-12-18 1c33327
Added more tests to testthat framework
