### 2016-12-18 4761260
Added all the necessary tests
