+++ tests/testthat/test-fusesma.sim.R
@@ -1,0 +1,145 @@
+context("test-fusesma.sim")
+
+test_that("Single run", {
+
+  data(fuse_hydrological_timeseries)
+  myDELTIM <- 1
+  myMID <- 60
+
+  set.seed(1)
+  parameters <- generateParameters(1)
+
+  x <- round(zoo::coredata(fusesma.sim(DATA = fuse_hydrological_timeseries,
+                                       mid = myMID,
+                                       deltim = myDELTIM,
+                                       rferr_add = parameters$rferr_add,
+                                       rferr_mlt = parameters$rferr_mlt,
+                                       frchzne = parameters$frchzne,
+                                       fracten = parameters$fracten,
+                                       maxwatr_1 = parameters$maxwatr_1,
+                                       percfrac = parameters$percfrac,
+                                       fprimqb = parameters$fprimqb,
+                                       qbrate_2a = parameters$qbrate_2a,
+                                       qbrate_2b = parameters$qbrate_2b,
+                                       qb_prms = parameters$qb_prms,
+                                       maxwatr_2 = parameters$maxwatr_2,
+                                       baserte = parameters$baserte,
+                                       rtfrac1 = parameters$rtfrac1,
+                                       percrte = parameters$percrte,
+                                       percexp = parameters$percexp,
+                                       sacpmlt = parameters$sacpmlt,
+                                       sacpexp = parameters$sacpexp,
+                                       iflwrte = parameters$iflwrte,
+                                       axv_bexp = parameters$axv_bexp,
+                                       sareamax = parameters$sareamax,
+                                       loglamb = parameters$loglamb,
+                                       tishape = parameters$tishape,
+                                       qb_powr = parameters$qb_powr)), 3)
+
+  y <- c(0, 0, 0, 0, 0.001, 0, 0, 0, 0, 0, 0.001, 0, 0, 0, 0, 0, 0.001,
+         0, 0, 0.001, 0, 0.001, 0, 0.001, 0, 0.002, 0, 0.046, 0.005, 0,
+         0.034, 0.312, 0.001, 0, 0, 0, 0, 0, 0.002, 0, 0.002, 0.002, 0.109,
+         0.01, 0.003, 0.002, 0.003, 0.053, 0.003, 0.033, 0.005, 0.009,
[…]
