+++ tests/testthat/test-fuserouting.sim.R
@@ -1,0 +1,234 @@
+context("test-fuserouting.sim")
+
+test_that("multiplication works", {
+
+  y <- c(0, 0, 0, 0, 0.001, 0, 0, 0, 0, 0, 0.001, 0, 0, 0, 0, 0, 0.001,
+         0, 0, 0.001, 0, 0.001, 0, 0.001, 0, 0.002, 0, 0.046, 0.005, 0,
+         0.034, 0.312, 0.001, 0, 0, 0, 0, 0, 0.002, 0, 0.002, 0.002, 0.109,
+         0.01, 0.003, 0.002, 0.003, 0.053, 0.003, 0.033, 0.005, 0.009,
+         0.006, 0.085, 0.001, 0, 0, 0.001, 0.003, 0.037, 0.031, 0.006,
+         0, 0.01, 0.007, 0.008, 0.003, 0.006, 0.001, 0.006, 0.005, 0,
+         0.001, 0.002, 0.008, 0.009, 0.025, 0.008, 0.001, 0, 0, 0, 0,
+         0.007, 0.012, 0.019, 0.008, 0.014, 0, 0.006, 0.01, 0, 0, 0, 0,
+         0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0.003, 0.018, 0.006, 0.001,
+         0.001, 0.002, 0.019, 0.006, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+         0, 0, 0, 0, 0, 0, 0, 0, 0, 0.01, 0, 0.027, 0.001, 0.007, 0, 0.002,
+         0, 0.001, 0, 0.018, 0.001, 0.013, 0.005, 0.011, 0.004, 0, 0.002,
+         0.003, 0.009, 0.01, 0.006, 0.001, 0.001, 0.002, 0, 0, 0.001,
+         0, 0, 0, 0, 0, 0.006, 0.002, 0.001, 0.001, 0, 0, 0, 0, 0, 0,
+         0, 0, 0, 0, 0, 0, 0, 0.001, 0, 0, 0.001, 0, 0, 0, 0, 0.021, 0,
+         0.003, 0.01, 0.002, 0.023, 0.018, 0.003, 0, 0.007, 0, 0, 0.002,
+         0.001, 0, 0, 0, 0, 0, 0.001, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+         0, 0.012, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0.001, 0, 0.005, 0.001,
+         0.004, 0.002, 0.001, 0.001, 0, 0, 0.013, 0, 0.005, 0.001, 0,
+         0.011, 0.026, 0, 0, 0.031, 0.001, 0.006, 0.001, 0, 0, 0, 0.006,
+         0, 0, 0, 0, 0, 0.022, 0, 0.017, 0.007, 0.006, 0.003, 0, 0.01,
+         0.002, 0.001, 0.011, 0.002, 0.017, 0.005, 0.006, 0, 0, 0, 0,
+         0, 0, 0, 0.001, 0.009, 0.003, 0.002, 0.004, 0, 0, 0, 0, 0.016,
+         0, 0.006, 0.002, 0, 0, 0, 0.001, 0, 0, 0, 0, 0, 0.001, 0.001,
+         0.002, 0.003, 0.008, 0.03, 0, 0, 0.01, 0.026, 0, 0, 0, 0, 0.002,
+         0.013, 0.001, 0.002, 0.011, 0.005, 0, 0.001, 0.003, 0, 0, 0.001,
[…]
