### 2016-12-18 cvitolo/fuse#13 @cvitolo
You right, I have added all the necessary tests.
