### 2016-12-18 7527c64
Updated documentation
