### 2016-12-18 0222863
Added description of package (as in paper) to README and vignette
