### 2016-12-18 e0873c2
Added a statement in the README saying code contributions are welcome.
