### 2016-12-18 #52 @cvitolo
@masalmon Many thanks for your thourough review, your suggestions greaty improved the consistency of the package.

I'm planning to update the release when the review is finished.

I have added the description of the package (as in the paper) to the README and vignette as you suggested.

I have added many more tests and addressed all the issues.

I have added info on related packages and future developments to the paper.

I have added a statement in the README saying code contributions are welcome.
