### 2016-12-18 cvitolo/fuse#11 @maelle
But you still use a random seed before calling them, right? Not important though :-)
