### 2016-12-18 cvitolo/fuse#13 @maelle
Ok, great, I see the tests, but I don't understand why the coverage as indicated [here](https://codecov.io/github/cvitolo/fuse?branch=master) doesn't seem to have changed? If you run covr locally, it must be a bit higher, right?
