### 2016-12-18 #52 @masalmon
Looks good to me, thanks for your work @cvitolo ! 👏

@arfon I'd recommend accepting the paper now 😀
