### 2016-12-18 cvitolo/fuse#16 @maelle
One test is failing
This issue is part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/52)

I actually get an error now when running the tests,
```
1. Failure: Ensemble run (@test-fuse.R#54) -------------------------------------
all(discharges == y) not equal to TRUE.
1 element mismatch

```
