### 2016-12-18 cvitolo/fuse#16 @maelle
Hmm, now it doesn't fail, installation was probably the issue, sorry!
