### 2016-12-18 cvitolo/fuse#13 @cvitolo
Yes, that's weird. I don't know why that happens. If I run covr locally I get a slightly higher percentage: ~63.46%. I guess it considers also the source code in the src folder....
