### 2016-12-18 cvitolo/fuse#16 @cvitolo
I'm glad it is not failing! phewwww
