### 2016-12-18 cvitolo/fuse#11 @cvitolo
Sorry, I should have clarified that I only set the seed before running `generateParameters` to have always the same parameters in my examples.
