### 2016-12-20 cfeb780
Updated doi in rearms.rmd
