### 2016-12-20 #52 @cvitolo
@arfon I just made a new release (v3.2), here is the new doi: http://doi.org/10.5281/zenodo.212822
